% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.factor.R
\name{conv.factor}
\alias{conv.factor}
\title{Convert variables to factors}
\usage{
conv.factor(data, convNames = NULL, exceptNames = NULL, uniqueNum = 5)
}
\arguments{
\item{data}{Input data frame.}

\item{convNames}{Names of variable to convert, the default is
\code{convNames = NULL}.}

\item{exceptNames}{Names of variables to be excluded from conversion, the
default is \code{convNames = NULL}.}

\item{uniqueNum}{Variables of less than or equal to a specific number of
unique values in the  to be converted to factors, the default is
\code{uniqueNum = 5}.}
}
\value{
A data frame of converted variables.
}
\description{
Convert variables to factors
}
\examples{
nhanes.fix <- conv.factor(data = nhanes, convNames = c("age", "hyp"))

}
