% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.mcar.R
\name{gen.mcar}
\alias{gen.mcar}
\title{Generate missing cells in the dataset}
\usage{
gen.mcar(df, prop.na = 0.2, warn.empty.row = TRUE, ...)
}
\arguments{
\item{df}{Input complete data frame}

\item{prop.na}{Proportion of missing cells, default = 0.2}
}
\value{
A data frame containing NAs
}
\description{
Generate missing cells in the dataset
}
