% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.rfnode.prox.R
\name{imp.rfnode.prox}
\alias{imp.rfnode.prox}
\title{Multiple imputation using chained random forests and node proximities}
\usage{
imp.rfnode.prox(
  data,
  num.imp = 5,
  max.iter = 5,
  num.trees = 10,
  pre.boot = TRUE,
  print.flag = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data. Missing
values should be coded as \code{NA}s.}

\item{num.imp}{Number of multiple imputations. The default is 5.}

\item{max.iter}{Number of iterations. The default is 5.}

\item{num.trees}{Number of trees to build, default to \code{10}.}

\item{pre.boot}{Perform bootstrap prior to modeling to get 'proper'
multiple imputation, i.e. accommodating sampling variation in estimating
population regression parameters (see Shah et al. 2014).
It should be noted that if \code{TRUE}, this option is in effect even if the
number of trees is set to one.}

\item{print.flag}{If \code{TRUE}, details will be sent to console.}

\item{...}{Other arguments to pass down.}
}
\value{
An object of S3 class \code{mids}.
}
\description{
\code{RfEmpImp} multiple imputation method for mixed types of variables,
using conditional distributions formed by proximity measures of random
forests (both in-bag and out-of-bag observations will be included).
}
\details{
\code{imp.rfnode.prox} multiple imputation, for missing observations, the
non-missing observations used for imputation will be found by whether two
observations can be retrieved from the same predicting node, the observations
used for imputation may not be necessarily be contained in the node.
}
\examples{
\donttest{
nhanesFix <- nhanes
nhanesFix[, c("age", "hyp")] <- lapply(nhanes[, c("age", "hyp")], as.factor)
impRfEmpImp <- imp.rfnode.prox(nhanesFix, printFlag = FALSE)
}

}
\references{
Hong, Shangzhi, et al. "Multiple imputation using chained random forests"
arXiv:2004.14823.

Zhang, Haozhe, et al. "Random Forest Prediction Intervals."
The American Statistician (2019): 1-20.

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.

Malley, James D., et al. "Probability machines." Methods of information
in medicine 51.01 (2012): 74-81.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
