% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.rfemp.R,
%   R/mice.impute.rfemp.cont.R, R/mice.impute.rfemp.cate.R
\name{mice.impute.rfemp}
\alias{mice.impute.rfemp}
\alias{mice.impute.rfemp.cont}
\alias{mice.impute.rfemp.cate}
\title{Random Forest EMPowered IMPutation}
\usage{
mice.impute.rfemp(
  y,
  ry,
  x,
  wy = NULL,
  num.trees = 10,
  num.trees.cont = NULL,
  sym.cont = TRUE,
  num.trees.cate = NULL,
  use.pred.prob.cate = TRUE,
  pre.boot = TRUE,
  emp.err.cont = TRUE,
  alpha.emp = 0,
  ...
)

mice.impute.rfemp.cont(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.cont = 10,
  sym.cont = TRUE,
  pre.boot = TRUE,
  emp.err.cont = TRUE,
  alpha.emp = 1,
  ...
)

mice.impute.rfemp.cate(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.cate = 10,
  use.pred.prob.cate = TRUE,
  pre.boot = TRUE,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{num.trees}{Number of trees to build, default to \code{10}; overrides
\code{num.trees.cont} and \code{num.trees.cate}}

\item{num.trees.cont}{Number of trees to build for continuous variables,
default to \code{10}}

\item{sym.cont}{Logical, \code{TRUE} for assuming symmetric empirical error
distribution, \code{FALSE} for asymmetric empirical error distribution,
default to \code{TRUE}. This option is ignored when \code{emp.err.cont} is
set to \code{FALSE}.}

\item{num.trees.cate}{Number of trees to build for categorical variables,
default to \code{10}}

\item{use.pred.prob.cate}{Logical, \code{TRUE} for assigning categories based
on predicted probabilities, \code{FALSE} for imputation based on majority
votes, default to \code{TRUE}}

\item{pre.boot}{Perform bootstrap prior to imputation to get 'proper'
imputation, i.e. accommodating sampling variation in estimating population
regression parameters (see Shah et al. 2014)}

\item{emp.err.cont}{Logical, \code{TRUE} for using
empirical error; \code{FALSE}, for assuming normal distribution for the
prediction error, the variance equals to overall out-of-bag prediction error,
i.e. mean squared error (see Shah et al. 2014).}

\item{alpha.emp}{The "significance level" for empirical error
distribution in prevention for outliers.
Set alpha = 0.05 to use 95\% confidence level for empirical error
distribution. Default is \code{0.0}, and the empirical error distribution is
kept intact.}

\item{...}{Other arguments to pass down}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
\code{RfEmpImp} imputation methods, adapter for \code{mice} samplers.
These functions can be called by the \code{mice} sampler function.

\code{mice.impute.rfemp} is for mixed types of variables, and calls
corresponding functions based on variable types. Categorical variables
should be of type \code{factor} or \code{logical}.

\code{mice.impute.rfemp.cont} is for continuous variables, performing
imputation based on the empirical distribution of out-of-bag
prediction errors of random forests.

\code{mice.impute.rfemp.cate} is for mixed categorical variables,
performing imputation based on predicted probabilities.
}
\details{
\code{RfEmpImp} Imputation sampler for: continuous variables based on the
empirical distribution of out-of-bag prediction errors of random forests;
categorical variables based on probability machines
}
\examples{
\donttest{
impRfEmpImp <- mice(nhanes, method = "rfempimp", m = 10,
max.iter = 10, maxcor = 1.0, printFlag = FALSE)#' }

}
\references{
Zhang, Haozhe, et al. "Random Forest Prediction Intervals."
The American Statistician just-accepted (2019): 1-20.

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.

Malley, James D., et al. "Probability machines." Methods of information
in medicine 51.01 (2012): 74-81.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
