% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.rfempimp.continuous.R
\name{mice.impute.rfempimp.continuous}
\alias{mice.impute.rfempimp.continuous}
\title{Multiple imputation  based on empirical distribution for continuous
variables, sampler adapter for MICE}
\usage{
mice.impute.rfempimp.continuous(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.continuous = 10,
  symmetry.continuous = TRUE,
  pre.bootstrap = TRUE,
  empirical.error.continuous = TRUE,
  alpha.empirical = 1,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{num.trees.continuous}{Number of trees to build for imputing continuous
variables, default to \code{10}}

\item{symmetry.continuous}{Logical indicator, \code{TRUE} for
assuming symmetry error distribution, \code{FALSE}, default to \code{FALSE}}

\item{pre.bootstrap}{Perform bootstrap prior to imputation to get 'proper'
imputation, i.e. accommodating sampling variation in estimating population
regression parameters (see Shah et al. 2014).}

\item{empirical.error.continuous}{Logical indicator, \code{TRUE} for using
empirical error; \code{FALSE}, for assuming normal distribution for error,
the variance equals to overall out-of-bag prediction error,
i.e. mean squared error (see Shah et al. 2014).}

\item{alpha.empirical}{The "significance level" for empirical error
distribution in prevention for outliers.
Set alpha = 0.05 to use 95\% confidence level for empirical error
distribution. Default is \code{0.0}, and the empirical error distribution is
kept intact.}

\item{...}{Other arguments to pass down}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Multiple imputation  based on empirical distribution for continuous
variables, sampler adapter for MICE
}
\references{
Zhang, Haozhe, et al. "Random Forest Prediction Intervals."
The American Statistician (2019)

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
