% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.rfnode.R,
%   R/mice.impute.rfnode.cond.R, R/mice.impute.rfnode.prox.R
\name{mice.impute.rfnode}
\alias{mice.impute.rfnode}
\alias{mice.impute.rfnode.cond}
\alias{mice.impute.rfnode.prox}
\title{Multiple imputation based on predicting nodes of random forest}
\usage{
mice.impute.rfnode(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.node = 10,
  pre.boot = TRUE,
  use.node.cond.dist = TRUE,
  obs.eq.prob = FALSE,
  do.sample = TRUE,
  ...
)

mice.impute.rfnode.cond(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.node = 10,
  pre.boot = TRUE,
  obs.eq.prob = FALSE,
  ...
)

mice.impute.rfnode.prox(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.node = 10,
  pre.boot = TRUE,
  obs.eq.prob = FALSE,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{num.trees.node}{Number of trees to build for imputing continuous
variables, default to \code{10}}

\item{pre.boot}{Perform bootstrap prior to imputation to get 'proper'
imputation, i.e. accommodating sampling variation in estimating population
regression parameters (see Shah et al. 2014).}

\item{use.node.cond.dist}{If \code{TRUE}, use conditional distribution formed
by predicting nodes of random forest (out-of-bag observations were excluded);
if \code{FALSE}, use proximity-based imputation.}

\item{obs.eq.prob}{If \code{TRUE}, the candidate observations will be sampled
with equal probability.}

\item{do.sample}{If \code{TRUE}, draw samples for missing observations;
if \code{FALSE}, the corresponding observations numbers will be provided and
will cause errors for the sampler functions. For testing purposes only, and
use with caution.}

\item{...}{Other arguments to pass down}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
\code{RfNode} imputation methods, adapter for \code{mice} samplers.
These functions can be called by the \code{mice} sampler functions.

\code{mice.impute.rfnode} is the main function for performing imputation, and
both \code{mice.impute.rfnode.cond} and \code{mice.impute.rfnode.prox} call
this function. By default, \code{mice.impute.rfnode} works like
\code{mice.impute.rfnode.cond}.

\code{mice.impute.rfnode.cond} is for imputation using the conditional formed
by the predicting nodes of random forests. To use this function, set
\code{method = "rfnode.cond"}.

\code{mice.impute.rfnode.prox} is for imputation based on proximity measures
from random forests, and provides functionality similar to
\code{mice.impute.rf}. To use this function, set
\code{method = "rfnode.prox"}.
}
\details{
`RfNode` Imputation sampler for: continuous variables based on the empirical
distribution of out-of-bag prediction errors of random forests;
categorical variables based on probability machines
}
\examples{
\donttest{
# Using "rfnode.cond"
impRfNodeCond <- mice(nhanes, method = "rfnode.cond", m = 10,
max.iter = 10, maxcor = 1.0, printFlag = FALSE)

# Using "rfnode.prox"
impRfNodeProx <- mice(nhanes, method = "rfnode.prox", m = 10,
max.iter = 10, maxcor = 1.0, printFlag = FALSE)
}

}
\references{
Doove, Lisa L., Stef Van Buuren, and Elise Dusseldorp.
"Recursive partitioning for missing data imputation in the presence of
interaction effects."
Computational Statistics & Data Analysis 72 (2014): 92-104.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
