% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.rfpred.emp.R
\name{mice.impute.rfpred.emp}
\alias{mice.impute.rfpred.emp}
\title{Multiple imputation using chained random forests: RfPred.Emp}
\usage{
mice.impute.rfpred.emp(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.cont = 10,
  sym.cont = TRUE,
  pre.boot = TRUE,
  emp.err.cont = TRUE,
  alpha.emp = 0,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed.}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{num.trees.cont}{Number of trees to build for continuous variables,
default to \code{NULL} to use the value of \code{num.trees}.}

\item{sym.cont}{Logical, \code{TRUE} for assuming symmetric distribution of
empirical prediction errors, \code{FALSE} for asymmetric distribution of
empirical prediction errors, default to \code{TRUE}.
This option is invalid when \code{emp.err.cont} is set to \code{FALSE}.}

\item{pre.boot}{Perform bootstrap prior to imputation to get 'proper'
multiple imputation, i.e. accommodating sampling variation in estimating
population regression parameters (see Shah et al. 2014).
It should be noted that if \code{TRUE}, this option is in effect even if the
number of trees is set to one.}

\item{emp.err.cont}{Logical, \code{TRUE} for using
empirical error; \code{FALSE}, for assuming normal distribution for the
prediction error, the variance equals to overall out-of-bag prediction error,
i.e. mean squared error (see Shah et al. 2014).}

\item{alpha.emp}{The "significance level" for empirical distribution of
prediction errors, can be used for prevention for outliers (useful for highly
skewed variables). For example, set alpha = 0.05 to use 95\% confidence level
for empirical distribution of prediction errors.
Default is \code{0.0}, and the empirical error distribution is kept intact.}

\item{...}{Other arguments to pass down.}

\item{num.trees}{Number of trees to build, default to \code{10}.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}.
}
\description{
For continuous variables only.

This function is for \code{RfPred.Emp}
multiple imputation method, adapter for \code{mice} samplers.
In the \code{mice()} function, set \code{method = "rfpred.emp"} to call it.

The function performs multiple imputation based on the empirical distribution
of out-of-bag prediction errors of random forests.
}
\details{
\code{RfPred.Emp} imputation sampler.
}
\examples{
\donttest{
impRfPredEmp <- mice(nhanes, method = "rfpred.emp", m = 10,
max.iter = 10, maxcor = 1.0, printFlag = FALSE)#' }

}
\references{
Zhang, Haozhe, et al. "Random Forest Prediction Intervals."
The American Statistician (2019): 1-20.

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.

Malley, James D., et al. "Probability machines." Methods of information
in medicine 51.01 (2012): 74-81.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
