% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.rfpred.norm.R
\name{mice.impute.rfpred.norm}
\alias{mice.impute.rfpred.norm}
\title{Multiple imputation using chained random forests: RfPred.Norm}
\usage{
mice.impute.rfpred.norm(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.cont = 10,
  pre.boot = TRUE,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed.}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{num.trees.cont}{Number of trees to build for continuous variables,
default to \code{NULL} to use the value of \code{num.trees}.}

\item{pre.boot}{Perform bootstrap prior to imputation to get 'proper'
multiple imputation, i.e. accommodating sampling variation in estimating
population regression parameters (see Shah et al. 2014).
It should be noted that if \code{TRUE}, this option is in effect even if the
number of trees is set to one.}

\item{...}{Other arguments to pass down.}

\item{num.trees}{Number of trees to build, default to \code{10}.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}.
}
\description{
For continuous variables only.

This function is for \code{RfPred.Norm}
multiple imputation method, adapter for \code{mice} samplers.
In the \code{mice()} function, set \code{method = "rfpred.norm"} to call it.

The function performs multiple imputation based on normality assumption using
out-of-bag mean squared error as the estimate for variance.
}
\details{
\code{RfPred.Emp} imputation sampler.
}
\examples{
\donttest{
impRfPredEmp <- mice(nhanes, method = "rfpred.norm", m = 10,
max.iter = 10, maxcor = 1.0, printFlag = FALSE)#' }

}
\references{
Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.
}
\author{
Shangzhi Hong, Henry S. Lynn*
}
