% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangerCallerSafe.R
\name{rangerCallerSafe}
\alias{rangerCallerSafe}
\title{Remove unnecessary arguments for \code{ranger} function}
\usage{
rangerCallerSafe(...)
}
\arguments{
\item{...}{Parameters to pass down.}
}
\value{
Constructed \code{ranger} object.
}
\description{
This function serves as an workaround for ranger function.
}
