% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.ests.R
\name{reg.ests}
\alias{reg.ests}
\title{Get regression estimates for pooled object}
\usage{
reg.ests(obj, ...)
}
\arguments{
\item{obj}{Pooled object from function \code{pool()}.}

\item{...}{Other parameters to pass down.}
}
\value{
A data frame containing coefficient estimates and corresponding
confidence intervals.
}
\description{
Get the estimates with corresponding confidence intervals after pooling.
}
