% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.mcar.R
\name{gen.mcar}
\alias{gen.mcar}
\title{Make Missing Cells in a Data Frame}
\usage{
gen.mcar(df, propMiss = 0.2, warnEmptyRow = TRUE, maxTry = 10)
}
\arguments{
\item{df}{Input complete data frame}

\item{propMiss}{Proportion of missing cells, default = 0.2}

\item{warnEmptyRow}{Whether try to warn about empty rows}
}
\value{
A data frame containing NAs
}
\description{
Make Missing Cells in a Data Frame
}
