% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.mcar.R
\name{gen.mcar}
\alias{gen.mcar}
\title{Generate missing (completely at random) cells in the dataset}
\usage{
gen.mcar(df, prop.na = 0.2, warn.empty.row = TRUE, ...)
}
\arguments{
\item{df}{Input data frame.}

\item{prop.na}{Proportion of missing cells, default = 0.2.}

\item{warn.empty.row}{Show warnings if empty rows were present.}

\item{...}{Other parameters (will be ignored).}
}
\value{
A data frame containing generated missing values.
}
\description{
Generate missing (completely at random) cells in the dataset
}
\examples{
data("mtcars")
mtcars.mcar <- gen.mcar(mtcars, warn.empty.row = FALSE)

}
