% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.rfpred.cate.R,
%   R/mice.impute.rfemp.R
\name{mice.impute.rfemp}
\alias{mice.impute.rfemp}
\alias{mice.impute.rfpred.cate}
\title{Multiple imputation for categorical variables based on predictions
of random forest}
\usage{
mice.impute.rfpred.cate(
  y,
  ry,
  x,
  wy = NULL,
  num.trees.cate = 10,
  use.pred.prob.cate = TRUE,
  pre.boot = TRUE,
  ...
)

mice.impute.rfemp(
  y,
  ry,
  x,
  wy = NULL,
  num.trees = 10,
  num.trees.cont = NULL,
  emp.err.cont = TRUE,
  alpha.emp = 0,
  sym.cont = TRUE,
  num.trees.cate = NULL,
  use.pred.prob.cate = TRUE,
  pre.boot = TRUE,
  ...
)
}
\arguments{
\item{y}{Vector to be imputed.}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{num.trees.cate}{Number of trees to build for categorical variables,
default to \code{NULL} to use the value of \code{num.trees}.}

\item{use.pred.prob.cate}{Logical, \code{TRUE} for assigning categories based
on predicted probabilities, \code{FALSE} for imputation based on majority
votes, default to \code{TRUE}.}

\item{pre.boot}{Perform bootstrap prior to imputation to get 'proper'
multiple imputation, i.e. accommodating sampling variation in estimating
population regression parameters (see Shah et al. 2014).
It should be noted that if \code{TRUE}, this option is in effect even if the
number of trees is set to one.}

\item{...}{Other arguments to pass down.}

\item{num.trees}{Number of trees to build, default to \code{10}.}

\item{num.trees.cont}{Number of trees to build for continuous variables,
default to \code{NULL} to use the value of \code{num.trees}.}

\item{emp.err.cont}{Logical, \code{TRUE} for using
empirical error; \code{FALSE}, for assuming normal distribution for the
prediction error, the variance equals to overall out-of-bag prediction error,
i.e. mean squared error (see Shah et al. 2014).}

\item{alpha.emp}{The "significance level" for empirical distribution of
prediction errors, can be used for prevention for outliers (useful for highly
skewed variables). For example, set alpha = 0.05 to use 95\% confidence level
for empirical distribution of prediction errors.
Default is \code{0.0}, and the empirical error distribution is kept intact.}

\item{sym.cont}{Logical, \code{TRUE} for assuming symmetric distribution of
empirical prediction errors, \code{FALSE} for asymmetric distribution of
empirical prediction errors, default to \code{TRUE}.
This option is invalid when \code{emp.err.cont} is set to \code{FALSE}.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}

Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}.
}
\description{
Part of project \code{RfEmpImp}, the function \code{mice.impute.rfpred.cate}
is for mixed categorical variables, performing imputation based on predicted
probabilities.

\code{RfEmpImp} multiple imputation method, adapter for \code{mice} samplers.
These functions can be called by the \code{mice} sampler function. In the
\code{mice()} function, set \code{method = "rfemp"} to call it.

\code{mice.impute.rfemp} is for mixed types of variables, and calls
corresponding functions based on variable types. Categorical variables
should be of type \code{factor} or \code{logical}.

For continuous variables, \code{mice.impute.rfpred.emp} is called, performing
imputation based on the empirical distribution of out-of-bag
prediction errors of random forests.

For categorical variables, \code{mice.impute.rfpred.cate} is called,
performing imputation based on predicted probabilities.
}
\details{
\code{RfEmpImp} Imputation sampler for: continuous variables based on the
empirical distribution of out-of-bag prediction errors of random forests;
categorical variables based on probability machines

\code{RfEmpImp} imputation sampler, the \code{mice.impute.rfemp} calls
\code{mice.impute.rfpred.emp} if the variable is numeric,
otherwise it calls \code{mice.impute.rfpred.cate}.
}
\examples{
\donttest{
impRfPredCate <- mice(nhanes, method = "rfpred.cate", m = 10,
max.iter = 10, maxcor = 1.0, printFlag = FALSE)
}

\donttest{
impRfEmpImp <- mice(nhanes, method = "rfemp", m = 10,
max.iter = 10, maxcor = 1.0, printFlag = FALSE)
}

}
\references{
Zhang, Haozhe, et al. "Random Forest Prediction Intervals."
The American Statistician just-accepted (2019): 1-20.

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.

Malley, James D., et al. "Probability machines." Methods of information
in medicine 51.01 (2012): 74-81.

Zhang, Haozhe, et al. "Random Forest Prediction Intervals."
The American Statistician (2019): 1-20.

Shah, Anoop D., et al. "Comparison of random forest and parametric
imputation models for imputing missing data using MICE: a CALIBER study."
American journal of epidemiology 179.6 (2014): 764-774.

Malley, James D., et al. "Probability machines." Methods of information
in medicine 51.01 (2012): 74-81.
}
\author{
Shangzhi Hong, Henry S. Lynn*

Shangzhi Hong, Henry S. Lynn*
}
