% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pkg-package.R
\docType{package}
\name{RfEmpImp-package}
\alias{RfEmpImp}
\alias{RfEmpImp-package}
\title{RfEmpImp: Multiple imputation using chained random forests}
\description{
Adapters for MICE.
    Functions for performing multiple imputation based on random forests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shangzhi-hong/RfEmpImp}
  \item Report bugs at \url{https://github.com/shangzhi-hong/RfEmpImp/issues}
}

}
\author{
\strong{Maintainer}: Shangzhi Hong \email{shangzhi-hong@hotmail.com}

Authors:
\itemize{
  \item Shangzhi Hong
  \item Henry S. Lynn [thesis advisor]
}

}
\keyword{internal}
