% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pkg-package.R
\docType{package}
\name{RfEmpImp-package}
\alias{RfEmpImp}
\alias{RfEmpImp-package}
\title{RfEmpImp: Multiple Imputation using Random Forest Empirical Prediction Error Distribution}
\description{
Adapter for MICE.
    Tools for performing multiple imputation based on random forest empirical
    prediction error distribution, can handle both continuous and categorical
    variables.
}
\author{
\strong{Maintainer}: Shangzhi Hong \email{shangzhi-hong@hotmail.com} (\href{https://orcid.org/0000-0003-3843-7040}{ORCID})

Authors:
\itemize{
  \item Shangzhi Hong \email{shangzhi-hong@hotmail.com} (\href{https://orcid.org/0000-0003-3843-7040}{ORCID})
  \item Henry S. Lynn \email{hslynn@shmu.edu.cn} [thesis advisor]
}

}
\keyword{internal}
