% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pkg-package.R
\docType{package}
\name{RfEmpImp-package}
\alias{RfEmpImp}
\alias{RfEmpImp-package}
\title{RfEmpImp: Multiple Imputation using Chained Random Forests}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Functions for methods for multiple imputation using chained random
    forests. Implemented algorithms can handle missing data in both continuous
    and categorical variables by using prediction-based or node-based
    conditional distributions constructed using random forests. For
    prediction-based imputation, the method based on the empirical distribution
    of out-of-bag prediction errors of random forests and the method based on
    normality assumption are provided. For node-based imputation, the method
    based on the conditional distribution formed by predicting nodes of random
    forests and the method based on measures of proximities of random forests
    are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shangzhi-hong/RfEmpImp}
  \item Report bugs at \url{https://github.com/shangzhi-hong/RfEmpImp/issues}
}

}
\author{
\strong{Maintainer}: Shangzhi Hong \email{shangzhi-hong@hotmail.com}

Other contributors:
\itemize{
  \item Henry S. Lynn [thesis advisor]
}

}
\keyword{internal}
